refactor(today): split page loading and actions
